added tests for the views for each app
